chore: move most of the code to `__init__.py` instead
